Diligent/pyproject.toml
[build-system]
requires = ["hatchling>=1.26"]
build-backend = "hatchling.build"

[project]
name = "diligent-dd"
version = "0.1.0"
description = "Due diligence state management CLI"
readme = "README.md"
requires-python = ">=3.11"
license = "LicenseRef-BSL-1.1"
license-files = ["LICENSE"]
authors = [
    { name = "Bryce Masterson" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Financial and Insurance Industry",
    "Topic :: Office/Business",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1",
    "openpyxl>=3.1.5",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
diligent = "diligent.cli:cli"

[project.urls]
Repository = "https://github.com/bmasterson1/diligent"

[tool.hatch.build.targets.wheel]
packages = ["diligent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
